change logic to determine state of checkbox in header row
If there are 1000 rows and we have rendered 100 then select all will
only select 100. In this case, technically all rows in react-table are
selected but we should not show the checked state. It leads to bad UX.
Indermediate would be better.

Also, split the additions of row selection column into a new useMemo. This
ensures that if a row is selected we are only re-calculating the row
selection column and not all.